Move backup elements out of Editor
